+++ src/index.test.js
@@ -1,68 +1,82 @@
 import chai from 'chai';
-import { Store } from './index';
+import sinon from 'sinon';
+import Firebase from 'firebase';
+import Fluxbase from './index';
 
 chai.should();
+
+// Stub Firebase class
+sinon.stub(Firebase, 'app', class {
+    database() {
+        return {
+            ref: path => ({
+                on: () => undefined
+            })
+        };
+    }
+});
+sinon.stub(Firebase, 'initializeApp', () => {
+   return new Firebase.app();
+});
+// END Stub
+
+describe('Fluxbase', () => {
+   it('should not have any stores registered', () => {
+       Fluxbase._stores.should.be.empty;
+   }); 
+});
 
 describe('Create store', () => {
     var store;
     
     it('should create an uninitialized store object', () => {
-        store = new Store();
+        store = Fluxbase.createStore();
         
-        store.should.not.have.property('database');
-        store.should.not.have.property('ref');
-        store.initialized.should.be.flase;
-        store.handlers.should.be.an('array').that.is.empty;
+        store.should.not.have.property('_database');
+        store.should.not.have.property('ref').that.is.a.function;;
+        store._isInit.should.be.false;
+    });
+    
+     it('should create a store using an object literal', () => {
+        store = Fluxbase.createStore({});
+        
+        store.should.have.property('_database');
+        store.should.have.property('ref').that.is.a.function;
+        store._isInit.should.be.true;
+    });
+    
+    it('should create a store using an existing Firebase object', () => {
+        let app = Firebase.initializeApp({});
+        store = Fluxbase.createStore(app);
+        
+        store.should.have.property('_database');
+        store.should.have.property('ref');
+        store._isInit.should.be.true;
     });
 });
 
-describe('Action handlers', () => {
-    var store;
-    var handler = action => null;
-    
-    beforeEach(() => {
-        store = new Store();
-    });
-    
-    it('should register an action handler', () => {
-        store.register(handler);
-        store.handlers.should.be.an('array').that.has.lengthOf(1);
-        store.handlers.should.include(handler);
-    });
-    
-    it('should unregister an action handler', () => {
-        store.register(handler);
-        store.unregister(handler);
-        store.handlers.should.be.an('array').that.is.empty;
-        store.handlers.should.not.not.include(handler);
-    });
-});
-
-describe('Event dispatch', () => {
-    it('should throw an error on uninitialized store', () => {
-        var store = new Store();
-        
+describe('Action dispatch', () => {
+    it('should throw an error on action without type', () => {
         (() => {
-            store.dispatch(null);
+            Fluxbase.dispatch({});
         }).should.throw(Error);
     });
     
-    it('should call all registered handlers', () => {
-        var store = new Store();
-        store.initialized = true;
+    it('all Stores should emit dispatched event', () => {
+        let spy1 = sinon.spy();
+        let spy2 = sinon.spy();
         
-        var ids = [1, 2, 3, 4, 5];
+        let store1 = Fluxbase.createStore();
+        let store2 = Fluxbase.createStore();
         
-        function handlerFactory(id) {
-            return action => {ids = ids.filter(elem => elem !== id)};
-        }
+        store1.on('test', spy1);
+        store2.on('test', spy2);
         
-        ids.forEach(id => {
-            store.register(handlerFactory(id));
+        Fluxbase.dispatch({
+            type: 'test'
         });
         
-        store.dispatch(null);
-        
-        ids.should.be.empty;
+        spy1.called.should.be.true;
+        spy2.called.should.be.true;
     });
 });